refactor(auth): sign-in method chooser, all methods under sign-in/
- /administration/sign-in is now a chooser: magic-link as the primary filled
  action, "Přihlásit heslem" only when the break-glass flag is on
- move the password form to its own page /administration/sign-in/password
  (gated loader redirects to the chooser when disabled)
- move magic-link routes from auth/magic-link to sign-in/magic-link so every
  method lives under sign-in/ (verify path becomes
  /administration/sign-in/magic-link/verify)
- 2FA failures return to the password page instead of the chooser
- link build + all internal redirects/links updated to the new paths

Note: the GAS web app's CONFIG.verifyPathPrefix must match the new verify path
and be redeployed.
