+++ app/routes.ts
@@ -76,8 +76,14 @@ export default [
 
       // Non-authenticated routes
       layout('routes/administration/__layout-non-authenticated/route.tsx', [
-        // Sign in
+        // Sign in — method chooser
         route('sign-in', 'routes/administration/sign-in/route.tsx'),
+
+        // Break-glass password sign-in form
+        route(
+          'sign-in/password',
+          'routes/administration/sign-in/password/route.tsx',
+        ),
 
         // Break-glass password second factor (TOTP) entry step
         route(
@@ -87,14 +93,14 @@ export default [
 
         // Magic-link sign-in: request form
         route(
-          'auth/magic-link',
-          'routes/administration/auth/magic-link/route.tsx',
+          'sign-in/magic-link',
+          'routes/administration/sign-in/magic-link/route.tsx',
         ),
 
         // Magic-link sign-in: verify + confirm
         route(
-          'auth/magic-link/verify',
-          'routes/administration/auth/magic-link/verify/route.tsx',
+          'sign-in/magic-link/verify',
+          'routes/administration/sign-in/magic-link/verify/route.tsx',
         ),
       ]),
