+++ app/routes/administration/sign-in/magic-link/_loader.ts
