+++ app/routes/administration/sign-in/magic-link/_middleware.ts
