+++ app/routes/administration/sign-in/magic-link/_schema.ts
