+++ app/routes/administration/sign-in/password/_meta.ts
@@ -1,0 +1,9 @@
+import { createPageTitle } from '~/utils/create-page-title'
+
+import type { Route } from './+types/route'
+
+export const meta: Route.MetaFunction = () => {
+  const title = createPageTitle('Administrace - přihlášení heslem')
+
+  return [{ title }]
+}
